+++ app/Dockerfile
@@ -4,8 +4,7 @@ EXPOSE 5000
 
 WORKDIR /app
 
-COPY requirements.txt /app
+COPY . /app
 RUN pip install -r requirements.txt
 
-COPY app.py /app
 CMD python app.py
