+++ app/templates/bill.html
