+++ db/billingdb.sql
@@ -34,8 +34,8 @@ CREATE TABLE IF NOT EXISTS `Trucks` (
 --
 -- Dumping data
 --
-INSERT INTO Provider (`id`,`name`) VALUES ('ALL'), ('pro1'),
-(3, 'pro2');
+--  INSERT INTO Provider (`id`,`name`) VALUES ('ALL'),('pro1'),
+-- (3, 'pro2');
 
 INSERT INTO Rates (`product_id`, `rate`, `scope`) VALUES ('1', 2, 'ALL'),
 (2, 4, 'pro1');
